config file implementation (colors only, key bindings are to come)
git-svn-id: https://svn.code.sf.net/p/gopherus/code/trunk@256 767d67df-c2ba-418b-950d-eab06a56d38d

+++ readflin.h
@@ -1,14 +1,15 @@
 /*
  * This file is part of the Gopherus project.
- * Copyright (C) 2013-2019 Mateusz Viste
+ * Copyright (C) 2013-2022 Mateusz Viste
  */
 
 #ifndef READFLIN_H
 #define READFLIN_H
 
 /* reads a single line from file descriptor f and fills memory at b with it
- * (including trailing \n).
- * returns length of line (still including the \n terminator) or 0 on EOF */
-unsigned short readfline(char *b, unsigned short blen, FILE *f);
+ * (without the \n or \r\n line terminator).
+ * returns length of line + 1 or 0 on EOF or when line too long to fit in b.
+ * on success (result > 0) the b buffer is guaranteed to be nul-terminated.  */
+size_t readfline(char *b, size_t blen, FILE *f);
 
 #endif
